Adding the dialog inside an activity to avoid java call issues
