+++ WordPress/src/main/res/values/styles.xml
@@ -361,6 +361,16 @@
         <item name="android:windowIsTranslucent">true</item>
         <item name="android:windowAnimationStyle">@android:style/Animation</item>
     </style>
+    
+    <style name="WordPress.TransparentDialog" parent="WordPress.NoActionBar">
+        <item name="android:windowNoTitle">true</item>
+        <item name="android:windowBackground">@android:color/transparent</item>
+        <item name="android:colorBackgroundCacheHint">@null</item>
+        <item name="android:windowIsTranslucent">true</item>
+        <item name="android:windowAnimationStyle">@android:style/Animation.Dialog</item>
+        <item name="android:backgroundDimEnabled">true</item>
+        <item name="android:windowContentOverlay">@null</item>
+    </style>
 
     <style name="TabBarStyle.WordPress" parent="@android:style/Widget.Material.ActionBar.TabBar">
         <item name="android:showDividers">middle</item>
